Legger tilbake TaskStep-metoder i TaskWorker
